fix clang-tidy defects

+++ cvmfs/sync_item.cc
@@ -217,7 +217,8 @@ catalog::DirectoryEntryBase SyncItemNative::CreateBasicCatalogDirent() const {
   dirent.size_             = graft_size_ > -1 ? graft_size_ :
                              this->GetUnionStat().st_size;
   dirent.mtime_            = this->GetUnionStat().st_mtime;
-  dirent.mtime_ns_         = this->GetUnionStat().st_mtim.tv_nsec;
+  dirent.mtime_ns_         = static_cast<int32_t>(
+                               this->GetUnionStat().st_mtim.tv_nsec);
   dirent.checksum_         = this->GetContentHash();
   dirent.is_external_file_ = this->IsExternalData();
   dirent.is_direct_io_     = this->IsDirectIo();
